update list/copy cobra short docs to reflect clip default commands
By default, when run without any subcommands or args, `clip` will
execute the `list` command, and when run with 1 arg, `clip` will execute
the `copy` command.

This behavior is already documented in the README, but this should help
make it more obvious to the user how things work when running `clip
help`/`clip --help`.