copy: remove size check for reading from stdin
Turns out this causes issues with commands that need to retreive data
and/or do not output their results immediately, ie:

```
curl -4sL echo.tjhop.io | clip copy
```

would error out because `curl` hadn't gotten the data yet. This didn't
come up in testing, since I was invoking the binary with `go run main.go
copy` which takes a second or so to initialize things.

Realistically, I can't think of a _bad_ thing that'd really happen with
not checking the input size. If it's empty, it'll just clear out the
clipboard, which is sane to me.